fix unsigned-signed comparison warning

+++ src/moveit_opw_kinematics_plugin.cpp
@@ -115,7 +115,7 @@ bool MoveItOPWKinematicsPlugin::setRedundantJoints(const std::vector<unsigned in
     ROS_ERROR_NAMED("srv", "This group cannot have redundant joints");
     return false;
   }
-  if (redundant_joints.size() > num_possible_redundant_joints_)
+  if (redundant_joints.size() > static_cast<std::size_t>(num_possible_redundant_joints_))
   {
     ROS_ERROR_NAMED("srv", "This group can only have %d redundant joints", num_possible_redundant_joints_);
     return false;
